requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
Django==5.2.1
django-celery-beat==2.8.1
django-extensions==4.1
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
gevent==25.9.1
greenlet==3.3.2
gunicorn==25.1.0
h3==4.4.2
idna==3.11
kombu==5.5.3
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==3.0
PyJWT==2.9.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
rdp==0.8
redis==6.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.event==6.1
zope.interface==8.2
whitenoise==6.9.0